integrating random price service

+++ P02-VehiclesAPI/pricing-service-graphql/src/main/java/com/udacity/pricingservicegraphql/entity/Price.java
@@ -1,6 +1,7 @@
 package com.udacity.pricingservicegraphql.entity;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 
 @Entity
 public class Price {
@@ -8,11 +9,11 @@ public class Price {
     @GeneratedValue(strategy = GenerationType.AUTO)
     private Long id;
 
-    private Long price;
+    private BigDecimal price;
     private String currency;
     private Long vehicle_id;
 
-    public Price(Long id, Long price, String currency, Long vehicle_id) {
+    public Price(Long id, BigDecimal price, String currency, Long vehicle_id) {
         this.id = id;
         this.price = price;
         this.currency = currency;
@@ -31,11 +32,11 @@ public class Price {
         this.id = id;
     }
 
-    public Long getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(Long price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
